+++ pkg/query-service/app/integrations/thirdPartyApi/model.go
@@ -1,11 +1,11 @@
-package thirdPartApi
+package thirdPartyApi
 
 import v3 "go.signoz.io/signoz/pkg/query-service/model/v3"
 
 type ThirdPartyApis struct {
 	Start    int64             `json:"start"`
 	End      int64             `json:"end"`
-	ShowIP   bool              `json:"show_ip,omitempty"`
+	ShowIp   bool              `json:"show_ip,omitempty"`
 	Domain   int64             `json:"domain,omitempty"`
 	Endpoint string            `json:"endpoint,omitempty"`
 	Filters  v3.FilterSet      `json:"filters,omitempty"`
