+++ pkg/query-service/app/integrations/thirdPartyApi/translator_test.go
@@ -1,4 +1,4 @@
-package thirdPartApi
+package thirdPartyApi
 
 import (
 	"testing"
